requirements.txt
openai>=1.0.0
anthropic>=0.34.0
arxiv>=1.4.0
PyYAML>=6.0
requests>=2.28.0
networkx>=3.2.0
pyvis>=0.3.2
scikit-learn>=1.6.0
